Remove shared spec
