+++ spec/features/issue_pages/evidence_spec.rb
@@ -26,11 +26,5 @@ describe 'issue pages evidence tab' do
       visit project_issue_path(current_project, issue)
       click_link("Evidence #{issue.evidence.count}")
     end
-
-    let(:default_columns) { ['Label', 'Title'] }
-    let(:hidden_columns) { ['Created by'] }
-    let(:filter) { { keyword: issue.evidence.first.title, filter_count: 1 } }
-
-    it_behaves_like 'a DataTable'
   end
 end
